+++ src/python/DAS/web/das_server.py
@@ -25,31 +25,29 @@ from cherrypy import config as cpconfig
 # DAS modules
 from DAS.utils.das_config import das_readconfig
 from DAS.web.das_webmanager import DASWebManager
-from DAS.web.das_web import DASWebService
-from DAS.web.das_cache import DASCacheService
-from DAS.web.das_web_srv import DASWebService as DASNewWebService
-#from DAS.web.das_expert import DASExpertService
+from DAS.web.das_web_srv import DASWebService
 from cherrypy.process.plugins import PIDFile
 
 class Root(object):
     """
     DAS web server class.
     """
-    def __init__(self, model, config):
-        self.model  = model
+    def __init__(self, config):
+        if  not isinstance(config, dict):
+            raise Exception('Wrong config type')
         self.config = config
         self.auth   = None
         self.pid    = None
         
     def configure(self):
         """Configure server, CherryPy and the rest."""
-        config = self.config[self.model]
+        config = self.config['web_server']
         cpconfig["engine.autoreload_on"] = False
 
         cpconfig["server.environment"] = config.get("environment", "production")
         cpconfig["server.thread_pool"] = int(config.get("thread_pool", 30))
         cpconfig["server.socket_port"] = int(config.get("port", 8080))
-        self.pid = config.get('pid', '/tmp/das_%s.pid' % self.model)
+        self.pid = config.get('pid', '/tmp/das_web.pid')
 
         cpconfig["server.socket_host"] = config.get("host", "0.0.0.0")
         cpconfig["server.socket_queue_size"] = \
@@ -70,49 +68,26 @@ class Root(object):
                           'tools.encode.on': True,
                           'tools.proxy.on': True,
                           'tools.gzip.on': True,
-#                          'tools.trailing_slash.on': False,
                           })
 
         log("loading config: %s" % cpconfig, 
-                                   context=self.model, 
+                                   context='web', 
                                    severity=logging.DEBUG, 
                                    traceback=False)
 
-    def start(self, server, blocking=True):
+    def start(self, blocking=True):
         """Configure and start the server."""
         self.configure()
         url_base = self.config['web_server']['url_base']
-        if  server == 'cache':
-            config = self.config.get('cache_server', {})
-            config['engine'] = engine
-            obj = DASCacheService(config) # mount cache server
-            tree.mount(obj, '/')
-        elif server == 'web':
-            # web server
-            config = self.config.get('web_server', {})
-            config['engine'] = engine
-            obj = DASWebService(config)
-            tree.mount(obj, url_base) # mount web server
-            # expert part
-#            config = self.config.get('expert_server', {})
-#            obj = DASExpertService(config)
-#            url = url_base + '/expert'
-#            tree.mount(obj, url) # mount expert server
-        elif server == 'new':
-            # new web server
-            config = self.config.get('web_server', {})
-            config['engine'] = engine
-            obj = DASNewWebService(config)
-            tree.mount(obj, url_base) # mount web server
-        else:
-            obj = DASWebManager({}) # pass empty config dict
-            tree.mount(obj, '/')
-
-        print "### %s, PID=%s" % (self.model, self.pid)
+        config = self.config.get('web_server', {})
+        config['engine'] = engine
+        obj = DASWebService(config)
+        tree.mount(obj, url_base) # mount web server
+        print "### DAS web server, PID=%s" % self.pid
         print pformat(tree.apps)
+        print pformat(config)
         pid = PIDFile(engine, self.pid)
         pid.subscribe()
-
         engine.start()
         if  blocking:
             engine.block()
@@ -124,8 +99,6 @@ def main():
     parser  = OptionParser()
     parser.add_option("-c", "--config", dest="config", default=False,
         help="provide cherrypy configuration file")
-    parser.add_option("-s", "--server", dest="server", default=None,
-        help="specify DAS server, e.g. web or cache")
     opts, _ = parser.parse_args()
 
     config  = das_readconfig()
@@ -134,22 +107,9 @@ def main():
         config = yaml.load(fdesc.read())
         fdesc.close()
 
-    # Choose which DAS server to start
-    if  opts.server == 'cache':
-        model = "cache_server"
-    elif opts.server == 'web':
-        model = "web_server"
-    elif opts.server == 'new':
-        model = "web_server"
-    else:
-        print "Please specify which DAS server you want to start, see --help"
-        sys.exit(1)
-
     # Start DAS server
-    print "\n### Start %s server with DAS configuration:" % model
-    print pformat(config)
-    root = Root(model, config)
-    root.start(opts.server)
+    root = Root(config)
+    root.start()
 
 if __name__ == "__main__":
     main()